chore: move to v2 pg submodule
