+++ src/api/routes/v2/blocks.ts
@@ -31,7 +31,7 @@ export function createV2BlocksRouter(db: PgStore): express.Router {
       if (!validRequestQuery(req, res, CompiledBlocksQueryParams)) return;
       const query = req.query as BlocksQueryParams;
 
-      const { limit, offset, results, total } = await db.getV2Blocks(query);
+      const { limit, offset, results, total } = await db.v2.getBlocks(query);
       const response: NakamotoBlockListResponse = {
         limit,
         offset,
@@ -50,7 +50,7 @@ export function createV2BlocksRouter(db: PgStore): express.Router {
       if (!validRequestParams(req, res, CompiledBlockParams)) return;
       const params = req.params as BlockParams;
 
-      const block = await db.getV2Block(params);
+      const block = await db.v2.getBlock(params);
       if (!block) {
         res.status(404).json({ errors: 'Not found' });
         return;
@@ -73,7 +73,7 @@ export function createV2BlocksRouter(db: PgStore): express.Router {
       const query = req.query as TransactionPaginationQueryParams;
 
       try {
-        const { limit, offset, results, total } = await db.getV2BlockTransactions({
+        const { limit, offset, results, total } = await db.v2.getBlockTransactions({
           ...params,
           ...query,
         });
